Extracts method to it's own file on challenge 2, and adds a solution for challenge 5
