+++ challenge_2/ruby/selectivealso/methods.rb
@@ -1,0 +1,12 @@
+class Array
+  def find_unique
+    count = Hash.new(0)
+    
+    self.each do |num|
+      count[num] += 1
+    end
+
+    return count.key(1)
+
+  end
+end
